+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 ~~~~~~~~~
 
+0.36.0 (April 2024)
+-------------------
+
+* Updated pytket version requirement to 1.27.
+* Fix incorrect readout when circuit contains implicit swaps 
+
 0.35.0 (March 2024)
 -------------------